+++ packages/shared/sdk-client/src/api/index.ts
@@ -2,5 +2,6 @@ import ConnectionMode from './ConnectionMode';
 
 export * from './LDOptions';
 export * from './LDClient';
+export * from './LDEvaluationDetail';
 
 export { ConnectionMode };