+++ packages/shared/sdk-client/src/configuration/validators.ts
@@ -1,18 +1,7 @@
 // eslint-disable-next-line max-classes-per-file
-import { noop, TypeValidator, TypeValidators } from '@launchdarkly/js-sdk-common';
+import { TypeValidator, TypeValidators } from '@launchdarkly/js-sdk-common';
 
 import { type LDOptions } from '../api';
-import { LDInspection } from '../api/LDInspection';
-
-class BootStrapValidator implements TypeValidator {
-  is(u: unknown): boolean {
-    return typeof u === 'object' || typeof u === 'undefined' || u === null;
-  }
-
-  getType(): string {
-    return `LDFlagSet`;
-  }
-}
 
 class ConnectionModeValidator implements TypeValidator {
   is(u: unknown): boolean {
@@ -46,22 +35,11 @@ const validators: Record<keyof LDOptions, TypeValidator> = {
 
   pollInterval: TypeValidators.numberWithMin(30),
 
-  // TODO: inspectors
-  // @ts-ignore
-  inspectors: TypeValidators.createTypeArray<LDInspection>('LDInspection[]', {
-    type: 'flag-used',
-    method: noop,
-    name: '',
-  }),
   privateAttributes: TypeValidators.StringArray,
 
   applicationInfo: TypeValidators.Object,
-  // TODO: bootstrap
-  bootstrap: new BootStrapValidator(),
   wrapperName: TypeValidators.String,
   wrapperVersion: TypeValidators.String,
-  // TODO: hash
-  hash: TypeValidators.String,
 };
 
 export default validators;