+++ packages/shared/sdk-client/src/index.ts
@@ -1,8 +1,19 @@
 import LDClientImpl from './LDClientImpl';
 
-export * as platform from '@launchdarkly/js-sdk-common';
-export * from './api';
-
 export * from '@launchdarkly/js-sdk-common';
+
+export * as platform from '@launchdarkly/js-sdk-common';
+
+// To replace the exports from `export *` we need to name them.
+// So the below exports replace them with the Node specific variants.
+
+// These exports are explicit to override those from common.
+export type {
+  LDEvaluationDetail,
+  LDEvaluationDetailTyped,
+  LDClient,
+  LDOptions,
+  ConnectionMode,
+} from './api';
 
 export { LDClientImpl };